+++ Trebuchet/CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ---
 
 ## [1.41.0-alpha] - 2026-07-01
-**Branch**: `trebuchet/issue-2267` | **PR**: #TBD
+**Branch**: `trebuchet/issue-2267` | **PR**: #2643
 
 ### Feature: Create new HAK from scratch (#2267)